build: move strlcpy to libmissing
Problem: strlcpy() is provided by libc on BSD based systems,
but we assume it is never provided by the system.

Add it to libmissing so it will be conditionally compiled.
Make includes conditional on HAVE_STRLCPY.
