+++ src/common/libmissing/strlcpy.c
